+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Improved
 - added support for new Omniva XML response namespace
+- improved shipping methods restrictions check
 
 ## [1.11.0]
 ### Fixed